+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.2
+
+* Simplify onResume implementation
+* Now, all you have to do is just replace `Navigator.push()` or `Navigator.pushNamed()` with `push()` or `pushNamed()`
+* Add `resume.data` and `resume.source` to store data from a screen you have pushed
+
+## 1.0.1+1
+
+* Update README.md
+
 ## 1.0.1
 
 * Add github link